+++ app/main.py
@@ -3,19 +3,19 @@ import os
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from app.domains.classroom.controller.teacher_class_router import router as teacher_classroom_router
-from app.domains.progress.controller.teacher_router import router as teacher_student_view_router
+from app.domains.progress.controller.teacher_student_router import router as teacher_student_view_router
 from app.domains.classroom.controller.student_class_router import router as student_class_router
 from app.domains.instruction.controller.teacher_router import router as teacher_instruction_router
 from app.domains.instruction.controller.student_router import router as student_assignment_router, stage3_router as student_stage3_orchestration_router
-from app.domains.content.controller.router import router as content_catalog_router
+from app.domains.content.controller.catalog_router import router as content_catalog_router
 from app.domains.content.stage1.router import router as student_stage1_router
 from app.domains.content.stage2.router import router as student_stage2_router
 from app.domains.content.stage3.router import router as student_stage3_router
-from app.domains.progress.controller.router import router as student_progress_router
-from app.domains.developer.router import router as system_router
-from app.domains.auth.controller.router import router as auth_router
+from app.domains.progress.controller.student_progress_router import router as student_progress_router
+from app.domains.developer.admin_router import router as system_router
+from app.domains.auth.controller.auth_router import router as auth_router
 from app.domains.auth.controller.admin_auth_router import router as admin_auth_router
-from app.domains.agent.controller.router import router as agent_router
+from app.domains.agent.controller.agent_router import router as agent_router
 from app.common.init.initialization import get_initialization_service
 from app.common.logging.logging_config import get_logger
