+++ tests/structure/test_domain_structure.py
@@ -7,16 +7,16 @@ def test_domain_packages_are_importable():
         "app.domains",
         "app.domains.developer",
         "app.domains.developer.indexing_service",
-        "app.domains.developer.router",
+        "app.domains.developer.admin_router",
         "app.domains.auth",
         "app.domains.auth.models",
         "app.domains.auth.repository.repository",
-        "app.domains.auth.controller.router",
+        "app.domains.auth.controller.auth_router",
         "app.domains.auth.schema.schemas",
         "app.domains.auth.service.service",
         "app.domains.classroom.controller.teacher_class_router",
         "app.domains.classroom.controller.student_class_router",
-        "app.domains.progress.controller.teacher_router",
+        "app.domains.progress.controller.teacher_student_router",
         "app.domains.classroom.dependency.dependencies",
         "app.domains.classroom.models",
         "app.domains.classroom.repository.repository",
@@ -32,14 +32,14 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.repository.repository",
         "app.domains.instruction.schema.schemas",
         "app.domains.instruction.service.service",
-        "app.domains.content.controller.router",
+        "app.domains.content.controller.catalog_router",
         "app.domains.content.stage1.router",
         "app.domains.content.stage1.schemas",
         "app.domains.content.stage1.service",
         "app.domains.content.stage2.router",
         "app.domains.content.stage2.schemas",
         "app.domains.content.stage2.service",
-        "app.domains.progress.controller.router",
+        "app.domains.progress.controller.student_progress_router",
         "app.domains.content.stage3.router",
         "app.domains.content",
         "app.domains.content.models",
@@ -53,7 +53,7 @@ def test_domain_packages_are_importable():
         "app.domains.content.stage3.schemas",
         "app.domains.content.stage3.service",
         "app.domains.agent",
-        "app.domains.agent.controller.router",
+        "app.domains.agent.controller.agent_router",
         "app.domains.agent.schema.schemas",
         "app.infrastructure.rag",
         "app.infrastructure.rag.retriever",
